+++ CHANGES.rst
@@ -9,6 +9,7 @@ Version 0.8.0 (UNRELEASED)
 - Adds job start and finish times to ``logs`` command output.
 - Adds possibility to filter by file name, size and last modified values to ``ls`` command.
 - Adds new command ``workspaces``.
+- Adds new ``--workspaces`` flag to ``validate`` command.
 - Adds the possibility to filter and display all the runs of a given workflow to the ``list`` command.
 - Changes ``delete`` to prevent workflow hard deletion.
 - Changes ``list`` to display interactive session status.